+++ src/controllers/DefaultController.php
@@ -22,7 +22,9 @@ use eseperio\filescatalog\actions\NewLinkAction;
 use eseperio\filescatalog\actions\PropertiesAction;
 use eseperio\filescatalog\actions\RemoveACL;
 use eseperio\filescatalog\actions\RenameAction;
+use eseperio\filescatalog\actions\SharedWithMe;
 use eseperio\filescatalog\actions\ShareViaEmail;
+use eseperio\filescatalog\actions\ShareWithUser;
 use eseperio\filescatalog\actions\UploadAction;
 use eseperio\filescatalog\actions\ViewAction;
 use eseperio\filescatalog\models\Directory;
@@ -51,12 +53,12 @@ class DefaultController extends \yii\web\Controller
                 'rules' => [
                     [
                         'allow' => true,
-                        'actions' => ['index', 'view', 'properties', 'download', 'bulk-download'],
+                        'actions' => ['index', 'view', 'properties', 'download', 'bulk-download', 'shared'],
                         'roles' => ['?', '@'],
                     ],
                     [
                         'allow' => true,
-                        'actions' => ['upload', 'new-folder', 'delete', 'bulk-delete', 'new-link', 'rename', 'email'],
+                        'actions' => ['upload', 'new-folder', 'delete', 'bulk-delete', 'new-link', 'rename', 'email','share'],
                         'roles' => ['@'],
                     ],
                     [
@@ -89,6 +91,11 @@ class DefaultController extends \yii\web\Controller
         ];
     }
 
+    /**
+     * @param $action
+     * @return bool
+     * @throws \yii\web\BadRequestHttpException
+     */
     public function beforeAction($action)
     {
         if (Yii::$app->request->isGet) {
@@ -124,9 +131,13 @@ class DefaultController extends \yii\web\Controller
             'inherit-acl' => ['class' => InheritAcl::class],
             'fake' => ['class' => FakeAction::class],
             'bulk-download' => ['class' => BulkDownload::class],
+            'shared' => ['class' => SharedWithMe::class],
         ];
-        if ($this->module->enableEmailSharing) {
+        if ($this->getModule()->enableEmailSharing) {
             $actions['email'] = ['class' => ShareViaEmail::class];
+        }
+        if ($this->getModule()->enableUserSharing) {
+            $actions['share'] = ['class' => ShareWithUser::class];
         }
         return $actions;