+++ src/models/InodeShare.php
@@ -1,0 +1,16 @@
+<?php
+
+
+namespace eseperio\filescatalog\models;
+
+use app\traits\ContentControlTrait;
+use eseperio\filescatalog\models\base\InodeShare as BaseInodeShare;
+use kartik\builder\Form;
+
+/**
+ * This model represents an inode shared between two users
+ */
+class InodeShare extends BaseInodeShare
+{
+
+}